requirements.txt
black==25.1.0
click==8.2.1
colorama==0.4.6
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
